Hoist cache.modify method declaration to ApolloCache. (#6362)
Background: https://github.com/apollographql/apollo-client/pull/6350#issuecomment-635666826

The modify method and its related types are fairly InMemoryCache-specific,
which is why it felt safest to confine cache.modify to InMemoryCache when
I brought it back in #6350. In practice, however, that confinement just
makes cache.modify harder to use (annoyingly, not helpfully) in cases
where an ApolloCache is all that's available.

You could also argue that cache.modify has roughly the same standing (in
terms of InMemoryCache-specificity) as cache.identify or cache.gc, which
are also part of the "Optional API" of the ApolloCache base class. In that
sense, cache.modify is simply joining those other AC3 APIs in ApolloCache
(and thus also InMemoryCache).
