+++ src/cache/inmemory/types.ts
@@ -1,15 +1,14 @@
-import { DocumentNode, FieldNode } from 'graphql';
+import { DocumentNode } from 'graphql';
 
 import { Transaction } from '../core/cache';
 import {
   StoreObject,
   StoreValue,
-  isReference,
   Reference,
 } from '../../utilities/graphql/storeUtils';
-import { FieldValueGetter, ToReferenceFunction } from './entityStore';
+import { FieldValueGetter } from './entityStore';
 import { KeyFieldsFunction } from './policies';
-import { SafeReadonly } from '../core/types/common';
+import { ToReferenceFunction, Modifier, Modifiers } from '../core/types/common';
 export { StoreObject, StoreValue, Reference }
 
 export interface IdGetterObj extends Object {
@@ -30,7 +29,7 @@ export interface NormalizedCache {
   has(dataId: string): boolean;
   get(dataId: string, fieldName: string): StoreValue;
   merge(dataId: string, incoming: StoreObject): void;
-  modify(dataId: string, modifiers: Modifier<any> | Modifiers): boolean;
+  modify(dataId: string, fields: Modifiers | Modifier<any>): boolean;
   delete(dataId: string, fieldName?: string): boolean;
   clear(): void;
 
@@ -106,43 +105,3 @@ export type CacheResolverMap = {
 // backwards compat
 export type CustomResolver = CacheResolver;
 export type CustomResolverMap = CacheResolverMap;
-
-export interface FieldSpecifier {
-  typename?: string;
-  fieldName: string;
-  field?: FieldNode;
-  args?: Record<string, any>;
-  variables?: Record<string, any>;
-}
-
-export interface ReadFieldOptions extends FieldSpecifier {
-  from?: StoreObject | Reference;
-}
-
-export interface ReadFieldFunction {
-  <V = StoreValue>(options: ReadFieldOptions): SafeReadonly<V> | undefined;
-  <V = StoreValue>(
-    fieldName: string,
-    from?: StoreObject | Reference,
-  ): SafeReadonly<V> | undefined;
-}
-
-export interface ModifyOptions {
-  id?: string;
-  fields: Modifiers | Modifier<any>;
-  optimistic?: boolean;
-  broadcast?: boolean;
-}
-
-export type Modifier<T> = (value: T, details: {
-  DELETE: any;
-  fieldName: string;
-  storeFieldName: string;
-  isReference: typeof isReference;
-  toReference: ToReferenceFunction;
-  readField: ReadFieldFunction;
-}) => T;
-
-export type Modifiers = {
-  [fieldName: string]: Modifier<any>;
-}
